#56 Fixed issue where post are automatically liked by the user.

+++ campfire/user/views.py
@@ -43,20 +43,26 @@ def feed(request):
 @login_required(login_url='/login')
 def post(request):
     post = Post.objects.get(id=request.GET['post_id'])
-    likes = Like.objects.filter(post_key=request.GET['post_id'])
-    is_Liked = False
+    print(Like.objects.filter(post_key=request.GET['post_id'], username=request.user.username))
     if Like.objects.filter(post_key=request.GET['post_id'], username=request.user.username).exists():
-        Like.objects.filter(post_key=request.GET['post_id'], username=request.user.username).delete()
-        is_Liked = False
-    else:
-        Like.objects.create(username=request.user.username, post_key=request.GET['post_id'])
         is_Liked = True
+    else:
+        is_Liked = False
+    if request.POST.get(' '):
+        is_Liked = False
+        if Like.objects.filter(post_key=request.GET['post_id'], username=request.user.username).exists():
+            Like.objects.filter(post_key=request.GET['post_id'], username=request.user.username).delete()
+            is_Liked = False
+        else:
+            Like.objects.create(username=request.user.username, post_key=request.GET['post_id'])
+            is_Liked = True
+
     comments = Comment.objects.filter(post_key=request.GET['post_id'])
     form = ModelFormForComment(request.POST, instance = Comment(username=request.user.username, post_key=request.GET['post_id']))
     if form.is_valid():
         form.save()
         return HttpResponseRedirect('/post/?post_id=' + request.GET['post_id'])
-    return render(request, 'post.html', {'MEDIA_URL': MEDIA_URL, 'post': post, 'likes': likes, 'is_Liked': is_Liked, 'comments': comments, 'form': form})
+    return render(request, 'post.html', {'MEDIA_URL': MEDIA_URL, 'post': post, 'is_Liked': is_Liked, 'comments': comments, 'form': form})
 
 
 @csrf_exempt
